pyproject.toml
[tool.poetry]
name = "review-analysis-ai"
version = "0.1.0"
description = ""
authors = ["yersmad <yersmad@gmail.com>"]
readme = "README.md"
packages = [{include = "review_analysis_ai"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.4"}
requests = "^2.31.0"
selenium = "^4.11.2"
streamlit = "^1.26.0"
langchain = "^0.0.281"
openai = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
